+++ Laravel/app/Http/Controllers/StaffController.php
@@ -3,10 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Models\Staff;
+use App\Models\User;
 use App\Models\Exhibition;
 use Illuminate\Http\Request;
 use App\Http\Requests\StaffRequest;
 use App\Traits\Paginatable;
+use Illuminate\Support\Facades\Hash;
 
 class StaffController extends Controller
 {
@@ -14,6 +16,12 @@ class StaffController extends Controller
     
     public function index()
     {
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
+        if(!$user || !$user->hasRole('admin')){
+            abort(403, 'Access denied!');
+        }
         
         $staff_query = Staff::with('exhibitions');
         $staff = $this->paginateWithPerPage($staff_query);
@@ -24,6 +32,13 @@ class StaffController extends Controller
     
     public function create()
     {
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
+        if(!$user || !$user->hasRole('admin')){
+            abort(403, 'Access denied!');
+        }
+
         $exhibitions = Exhibition::all();
         $isUpdate = false;
 
@@ -33,7 +48,26 @@ class StaffController extends Controller
 
     public function store(StaffRequest $request)
     {
-        $staff = Staff::create($request->validated());
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
+        if(!$user || !$user->hasRole('admin')){
+            abort(403, 'Access denied!');
+        }
+      
+        do {
+            $email = strtolower('worker' . uniqid() . '@gmail.com');
+        } while (User::where('email', $email)->exists());
+        
+        $user = User::create([
+            "name"=>$request->full_name,
+            "email"=>$email,
+            "password"=>Hash::make('123123'),
+            "role_id"=>2
+        ]);
+        
+        $staff = Staff::create([...$request->validated(), 'user_id' => $user->id]);
+
         $staff->exhibitions()->attach($request->exhibitions);
 
         return redirect()->route('staff.index')->with('success', 'Staff member created successfully');
@@ -42,13 +76,27 @@ class StaffController extends Controller
 
     public function show(Staff $staff)
     {
-        $staff->load('exhibitions');
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
+        if(!$user || !$user->hasRole('admin')){
+            abort(403, 'Access denied!');
+        }
+
+        $staff->load(['exhibitions', 'user']);
         return view('staff.show', compact('staff'));
     }
 
 
     public function edit(Staff $staff)
     {
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
+        if(!$user || !$user->hasRole('admin')){
+            abort(403, 'Access denied!');
+        }
+
         $exhibitions = Exhibition::all();
         $staff->load('exhibitions');
         $isUpdate = true;
@@ -57,6 +105,13 @@ class StaffController extends Controller
 
     public function update(StaffRequest $request, Staff $staff)
     {
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
+        if(!$user || !$user->hasRole('admin')){
+            abort(403, 'Access denied!');
+        }
+
         $staff->update($request->validated());
         $staff->exhibitions()->sync($request->exhibitions);
 
@@ -65,6 +120,13 @@ class StaffController extends Controller
 
     public function destroy(Staff $staff)
     {
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
+        if(!$user || !$user->hasRole('admin')){
+            abort(403, 'Access denied!');
+        }
+
         $staff->exhibitions()->detach();
         $staff->delete();
