+++ Laravel/app/Http/Controllers/TicketController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use App\Models\Ticket;
 use App\Models\Exhibition;
 use App\Models\Visitor;
-use Illuminate\Http\Request;
 use App\Http\Requests\TicketRequest;
 use App\Traits\Paginatable;
 
@@ -16,14 +15,23 @@ class TicketController extends Controller
 
     public function index()
     {
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
         $tickets_query = Ticket::query();
         $tickets = $this->paginateWithPerPage($tickets_query);
-
         return view('tickets.index', compact('tickets'));
     }
 
     public function create()
     {
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
+        if(!$user || !$user->hasRole('admin')){
+            abort(403, 'Access denied!');
+        }
+
         $exhibitions = Exhibition::all();
         $isUpdate = false;
 
@@ -32,6 +40,13 @@ class TicketController extends Controller
 
     public function store(TicketRequest $request)
     {
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
+        if(!$user || !$user->hasRole('admin')){
+            abort(403, 'Access denied!');
+        }
+
         $ticket = Ticket::create($request->validated());
         return redirect()->route('tickets.index')->with('success', 'Ticket created successfully');
     }
@@ -43,6 +58,13 @@ class TicketController extends Controller
 
     public function edit(Ticket $ticket)
     {
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
+        if(!$user || !$user->hasRole('admin')){
+            abort(403, 'Access denied!');
+        }
+
         $exhibitions = Exhibition::all();
         $isUpdate = true;
 
@@ -51,12 +73,26 @@ class TicketController extends Controller
 
     public function update(TicketRequest $request, Ticket $ticket)
     {
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
+        if(!$user || !$user->hasRole('admin')){
+            abort(403, 'Access denied!');
+        }
+
         $ticket->update($request->validated());
         return redirect()->route('tickets.index')->with('success', 'Ticket updated successfully.');
     }
 
     public function destroy(Ticket $ticket)
     {
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
+        if(!$user || !$user->hasRole('admin')){
+            abort(403, 'Access denied!');
+        }
+
         $ticket->delete();
 
         return redirect()->route('tickets.index')->with('success', 'Ticket deleted successfully!');
@@ -64,8 +100,14 @@ class TicketController extends Controller
 
     public function buy(Ticket $ticket)
     {
-        $visitor = Visitor::find(1); 
-        
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+        $id_user = auth()->id();
+
+        if(!$user){
+            abort(403, 'Access denied!');
+        }
+        $visitor = Visitor::where('user_id', $id_user)->first();
         $ticket->visitor()->associate($visitor);
         $ticket->save();
         
@@ -74,10 +116,16 @@ class TicketController extends Controller
 
     public function sell(Ticket $ticket)
     {
-        $id = $ticket->getVisitorId();
+        /** @var \App\Models\User $user */
+        $user = auth()->user();
+
+        if(!$user){
+            abort(403, 'Access denied!');
+        }
+
         $ticket->visitor()->dissociate();
         $ticket->save();
         
-        return redirect()->route('visitors.show', $id); 
+        return redirect()->route('profile.index')->with('success', 'Ticket sold successfully!'); 
     }
 }
